* simplify tests
